"11:50am. Since I got up this late, let me just see if I can find any transfer attribute videos.
https://youtu.be/TLwL1UTdtHg
[Tut] Field & Transfer Attribute node Explained - Blender Geometry Nodes 3.0 Field

Bradley had it.

Let me go for it.

https://youtu.be/2qWnRRkV9Zk
[Tut] 2 Uses of Capture Attribute Node Explained - Blender Geometry Nodes 3.0 Field

I got this in the video. I'll leave it until after I am finished.

https://youtu.be/TLwL1UTdtHg?t=352

Ah, so capture attribute does not transfer to different geometries.

I am going to have to play with this example a bit. After I've watched the video, I am going to try to replicate it without looking just to see whether I have understood.

https://youtu.be/TLwL1UTdtHg?t=532

Actually, I think I understand it completely. There is no real need to play with it. I'll just move to the next thing.

https://youtu.be/TLwL1UTdtHg?t=543

He says he is using the index mode. How is that useful?

12:05pm. Ok, I think I see it.

Also I've finally figured out what the viewer node is supposed to be doing.

12:10pm. I do not feel a need for it, but let me watch the capture attribute tutorial.

Hmmm, if I have points, even if endpoint selector failed me, could I just compare based on the index of it and use that as the selector?

https://youtu.be/2qWnRRkV9Zk?t=249

Ah, I can't forget about this node. I could very easily hack it with the mesh to points and make a mistake that way.

https://youtu.be/2qWnRRkV9Zk?t=288

I ran into this problem yesterday and figured it out.

https://youtu.be/2qWnRRkV9Zk?t=387

I really need to figure out what align euler to vector is doing.

https://youtu.be/X80M0q0gxIE
Quaternion or Euler, understanding rotation in Blender

I'll watch this after I am done.

https://youtu.be/2qWnRRkV9Zk?t=476

I wouldn't have thought mixing edge and face positions would work. How does that make sense?

Hmmm, why did he split the edges?

https://youtu.be/2qWnRRkV9Zk?t=538

Normals are associated with the face. I totally thought they were associated with the vertices.

12:30pm. Focus me, focus. Stop reading the liches thread.

https://youtu.be/X80M0q0gxIE
Quaternion or Euler, understanding rotation in Blender

Let me watch this.

https://youtu.be/0E1K9j9zoik
[Tut] Align Euler to Vector Node Explained - Blender Geometry Nodes 3.0 Field

I do not like that animation video, as I am not interested in amination. Let me watch this again. It is Bradley once more.

12:40pm. https://youtu.be/0E1K9j9zoik?t=183

This is rather instructive. Fine I'll subscribe to your channel.

I see now why normals can be converted to rotation using this. Because they just point in a certain direction.

12:45pm. https://youtu.be/0E1K9j9zoik?t=274

This is a good point, just what is rotation doing here?

Actually, does that connect to the rotation attribute for the geometry?

https://youtu.be/0E1K9j9zoik?t=316

Ohhh, I did not expect the factor to go into the negative. Yeah, the factor is not the factor between rotation and the vector. That can't be it.

12:50pm. https://youtu.be/0E1K9j9zoik?t=438

Not what I expected. I thought that the inputs should be all zeroes in order to make it point towards the origin...

Actually if you think that there is already rotation...

No way, there is no wait that is right.

Rather the normals themselves have to be pointing towards the origin.

If I put all zeroes as a vector it would not have any basis for rotation...

But didn't I assume that rotation is attached to the original geometry. Then it should work, right?

Ah, I'll play around with it later. Right now let me just watch it.

1pm. https://youtu.be/0E1K9j9zoik?t=690

I had no idea these constraints existed. This is ridiculous. The amount of things I do not know in Blender despite 3 full months of studying it is crazy.

https://youtu.be/0E1K9j9zoik?t=783

I didn't understand the previous part with two euler nodes, but I understand this.

https://youtu.be/57FaqP_Q36w
EVERY NODE in the Erindale Toolkit - Geometry Nodes

I won't watch this, but I'll keep the addon in mind. Let me move on.

1:10pm. By move on I mean have breakfast. This was a nice morning session.

Let's see, I really should play with subdivide edges in order to increase my understanding on how interpolating between faces and points works. Right now it does not make sense because there should be a different amount of them, so I do not know how Blender knows to decide between collections of different sizes. I need to play with that example. Maybe with align euler as well as I do not understand why adding that second vector makes a difference.

I should investigate that.

1:15pm. Sigh, I really want to start doing my own things, but I also want to increase my understanding. A few days ago I was complaining that the phyisics sims tutorials were extremely tedious, but I am findings these quite interesting. I am quite into this.

Well, I'll get through it at some point. There is not an infinite amount of tutorials to go through. Let me have breakfast here."