"10:40am. I was very restless tonight. I was so heated I had to remove all, but a single blanket in order to fall asleep. It took me a while before I got to that point.
My sleep might have been poor, but my will is high. Let me chill a bit and then I will start the Houdini tutorials. I really need to finish these basic ones before I can more on to the rest.

11:15am. Let me start instead of wasting so much time.

Time for the tutorials.

11:20am. https://youtu.be/NjtlYXRgqWc?list=PLhyeWJ40aDkUDHDOhZQ2UkCfNiQj7hS5W&t=137

Come to think of it, I never explicitly studied any animation tutorials. Ram's could count, but I never followed along, so this will be a first for me.

11:40am. I am starting to get used to the Houdini UI. The way it is designed is very good. If you press escape in the viewport, it is in move mode. But if you press enter, it goes into the transform node and you have those handles like in Blender. But it is not just that, it is actually sensitive to the node I have selected. Right now I am in the extrude node, and the transform handle allows me to extrude it.

https://youtu.be/NjtlYXRgqWc?list=PLhyeWJ40aDkUDHDOhZQ2UkCfNiQj7hS5W&t=671

Speaking as a programmer, this is a really good way of working.

12:10pm. https://youtu.be/NjtlYXRgqWc?list=PLhyeWJ40aDkUDHDOhZQ2UkCfNiQj7hS5W&t=954

I do like this. Houdini has it made. For Blender to have all of this, it would take an unimaginable amount of work.

https://youtu.be/NjtlYXRgqWc?list=PLhyeWJ40aDkUDHDOhZQ2UkCfNiQj7hS5W&t=1062

This kind of packing node would be impossible with Blender.

1:10pm. https://youtu.be/IcbzGNqnFkA?list=PLhyeWJ40aDkUDHDOhZQ2UkCfNiQj7hS5W&t=627

Let me stop here for the morning. I am overdue for breakfast.

1:40pm. Let me finish the ep of Kaiji Kaihatsu and then I'll go do the chores.

2:55pm. It is time to resume.

3pm. Let me do it. Time to do those shaders. Once I finish this series I can do more advanced tutorials.

3:20pm. This is strange. I noticed that in render view the background and platform shaders were not taking effect. I tried fiddling with them, and suddenly that happened.

https://youtu.be/IcbzGNqnFkA?list=PLhyeWJ40aDkUDHDOhZQ2UkCfNiQj7hS5W&t=855

I wonder how shadow intensity works. I thought shadows were only related to the amount of light in a scene. Strange.

3:35pm. https://www.quantamagazine.org/neural-noise-shows-the-uncertainty-of-our-memories-20220118/

Let me read this.

https://youtu.be/NHmAcO6Dob4?list=PLhyeWJ40aDkUDHDOhZQ2UkCfNiQj7hS5W
Houdini Isnt Scary Project - Part 3: Rigid Body Dynamics

After that it is rigid body dynamics for me.

3:40pm. Let me start.

https://youtu.be/NHmAcO6Dob4?list=PLhyeWJ40aDkUDHDOhZQ2UkCfNiQj7hS5W&t=264

To get it to start in the right place, I had to turn on use object transform. I wonder why it works for him despite that?

4:50pm. Time for part 4. Had to take a little break. Let me finish this.

5pm. Let me stop here for a bit, I have to do an extra bit of chores.

5:15pm. Let me resume.

5:25pm. And I am trying to do it on my own and weird things are happening. Is it possible that Houdini has caching bugs?

How do I clear it? Let me just watch the video. I do not know how to stop the platform from turning into water particles.

5:40pm. I've cleared up some of my confusion, but what is the point of that timeshift node? I had to have done something wrong since, it just results in an object that does nothing. Let me watch the video.

6:05pm. This is so annoying! I should have followed it step by step exactly instead of trying to explore on my own. I keep getting cache errors.

6:35pm. I am back from lunch.

Let me resume the tutorial.

7pm. Ugh, I have no idea what timeshift was supposed to be doing. And I really don't want to mess with IO nodes even though I should have. one thing I am dissapointed about is that he did not cover how to tranfer the velocity from the ball to the fluid.

7:05pm. Mhhh, damn it. I need to investigate the disk caching. I actually skipped the IO parts.

I also need to go back and figure out what that timeshift is doing.

7:10pm. I do not feel like it right now. As strange as it is, I couldn't finish the lecture series today despite how basic it is. I am going to do it tomorrow in its proper time and move on to making that soccer ball.

One step at a time. Let me have some rest here.

Despite not being able to do it, I was actually rather focused throughout. I think I internalized most of it. The problem is that there is so much of it. I can really see why Houdini is considered to have the hardest learning curve out of all the 3d programs out there. It is not just a 3d package like Blender, but a fully fledged programming environment. And on top of that it is massive, much more so than Blender. It feels like a program where if I could claim to understand 10% of it, I could call myself an expert.

I need to keep going forward with this. Maybe it seems like a waste, but assuming I could get an AI chip, maybe I could make it a side business to write physics simulation solvers for 3d programs. That woudl be a lot more realistic as a first step than trying to infer out intelligence algorithms via genetic programming. It would be excellent practice too as the intelligence algorithms will be some kind of phyisical system simulation anyway.

7:30pm> Let me close here, instead of stading here in a daze. I'll get this done tomorrow. At some point I will maange to establish my workflow. I need keep going forward until then. I could have already established it, but I will be able to reach a higher level of skill in the end if I add Houdini to my toolset. It will be worth it. I do not care about getting rich anymore.

I feel that artists doing it for the sake of art is a weird cope, but wanting to do ML for the sake either money or causing the Singularity did not make me any more successful. So I'll do this for the money, but I am not expecting anything great, not like with trading or ML. Just a little bit to start things off will be enough. Surely, I can at least get to the point where I am making 100$ a month. So far Spiral is making me 5$ a month, so I suppose that is something. I need to get more patrons. If I can get sponsors, I can actually start cultivating AI again. If not, then that is where the quest will end for me."