"11am. I am up. Today I am going to get some serious studying done if possible. I am going to watch the panel video from yesterday again and see if my focus has improved.
11:10am. Let me start. It is time I figure out how to sample different cables. Also how did he manage to do those bolts around them while having them fit? I need to focus.

I think I am about 10 days into Covid, and it should be time to get over it soon.

https://vimeo.com/364080941
Houdini Sci Fi Panel Generator | Part 1 | Modeling the Panel

I've watched this yesterday, but I could not understand a thing of it. This time I'll do better.

Yesterday I just about figured out how the for loop works. What I have to do now is seal the deal.

1:24 in. It is annoying that I do not have this particular node, nor do I know how to install them, but it does not matter. I'll leave figuring that out for later. Right now it feels like my focus is at a decent level. I am still not over it, but my head does not feel dizzy from the illness.

11:20am. I saw this yesterday, but did not have the energy to wonder about it properly so let me do that now. Just why is he duplicating the layer names?

Is he extracting the same layer except with a different brightness threeshold?

11:30am. This whole tutorial is so weird, I have no idea why it is structured the way it is.

11:40am. 15:17. I am getting good feedback from my subconscious. With the way things are, I can expect to get some real learning done today. That is the most important. Houdini might be big, but as long as I keep cracking at it, it will fall much like Blender.

11:55am. I am getting it. I thought about how I would have done it myself, and came to a conclussion that I could just use match size to push things around. But this way of doing things is not bad.

I am playing with fuse, and turning off the snap distance does exactly what I'd want without having to put in a large value. It merges all the points into a single value.

1:05pm. I've gone through the tutorial start to finish while paying careful attention to it.

This is it! I am back in action. This was an ideal study session. I've been paying attention and going through the steps myself. I've succeeded at learning a decent amount.

1:10pm. This is good enough. Let me have breakfast and do the chores. After that is done, I am going to start work on my own projects. It is finally time to get started on doing the pool scene in Houdini. All this automation capability is one thing, but I'll know how good it is once I try it with my own hands.

It is not a given that I have to use Houdini. Maybe I'll end up doing it all in Blender in the end.

1:50pm. Let me do the chores.

2:40pm. Done with that. By no means have I recovered. Maybe another week and this illness will blow over.

Before I start, let me take a look at how he did the preset.

Ah, I see.

2:50pm. It is time to get started. I opened the flower file for the first time in a while. Last time I did the bulb, and then I had trouble understanding how to do the stalk. But now that should not be a problem.

Let me get a refil and then I'll take stock of my condition.

2:55pm. Let me do a little review. Though I had a successful morning session, I am still not over the illness, so I won't force myself to work here.

Instead, let me plan things out.

I've been studying Houdini for 2.5 weeks now, so I know it well enough to form an opinion of it.

My view is that it is overrated. If it is merely distributing the flowers I could have used the scatter addon for Blender. the same might apply for buildings and things like that.

Despite all its sophistication, for doing things like sampling, Houdini strikes me as barely being better than Blender and I hate how I have to dip in a circular way using channels to make use of the switch node.

3:05pm. Yeah, Houdini is overrated. I think that in many places I'll try to automate things, and end up being slower than had I tried doing it by hand.

3:10pm. After doing the chores, the physical exhaustion is really hitting me.

Let me go back to the middle of last month. Back then, I had the option of going either with Blender or shoring up my knowledge even further and I choose the later.

Back then I had a really strong feeling of having grasped the basics. I feel that if I wanted to become an illustrator, back then I really could have done it.

I need to go back to that.

Basic modeling, basic sculpting, and then laying things out. Forget the advanced Houdini stuff. I am not going to be changing things that much.

3:30pm. Got the stalk done.

4:15pm. FFFFFFFFFFFFFFFFF...holy shit this is so annoying!

How do I do the correct orientation for the scattered points?

4:20pm. I am completely confused here.

4:30pm. I hate this. Why is the angle off for some leaves? When I try to do a cone, I get the right result.

https://www.youtube.com/results?search_query=houdini+scatter

Let me watch some of these.

I have no idea. I simply forgot how it works. This would not be a problem in Blender.

https://youtu.be/tCijuuNIRrs
New in Houdini 18.5 Pt. 5: Scatter & Align Plus Attribute From Pieces SOPs - The New Workflow

Let me watch this.

4:40pm. https://youtu.be/tCijuuNIRrs?t=110

This is a new thing.

5pm. I am not in my right mind. I am quite pissed off at Houdini right now.

Just what is the problem with the leaves? Do they have a random orientation or something?

5:15pm. https://youtu.be/N7CDHwgWKVo?t=406

Ah, here is the density for this. At any rate, I am going to have to make a simplified example to make sure that the orientation is right.

https://youtu.be/N7CDHwgWKVo?t=476

Let me stop here for the day.

5:20pm. I am not my usual self. Right now, I feel like smashing my fist through the monitor. I probably shouldn't be this pissed off over so little.

I still haven't gotten my senses of smell and taste back. I guess it too early to be making plans for getting back into the swing of things. Trying to pretend I am well is just pretense. Hard work is only something that happens when one is in proper health.

5:25pm. I am starting to form my view of things. Houdini might beat Blender in terms of workflow, but as a PL it has everything I most dislike about mainstream PLs. As a program it evolved solely by cramming shit into itself without a care for elegance. I thought that Blender's geometry nodes were complicated at first, but they are incredibly well designed compared to Houdini's bloat.

Houdini is the opposite of slim, it is an incredibly bloated piece of software.

5:30pm. I am so pissed off about this orientation thing. Houdini has normals, but then it has `up`, `orient`, `xform` and who knows what fucking else.

I am faced with this complexity, but I feel so dull and uninspired.

I got the stalk done. I certainly did that for the day. But I need to go further.

I need to figure out why the leaves have random rotation in places.

5:40pm. You know what I hate about Houdini? How unlike in Blender it does not rotate around an object, but hangs around freeform.

5:45pm. I am copying to points on a sphere, and I see that for the leaves to have proper orientation, I need to set `@up = {0,1,0};`. That gives it proper orientation.

This is what I've been suspecting is the problem. The points probably have the normals in the right place, but that still leaves a 360 degree of freedom with regards to what the other axis should be. In Blender this wasn't a problem for some reason.

Still, I find it strange that I have to set this manually using the attribute wrangle.

6:10pm. No, setting the up does not do jack when combined with the scatter and align node.

I am just bashing my head against the wall here.

Enough.

6:15pm. Sometimes just stopping when things aren't going your way is a sign of discipline. Maybe my mental acuity will be better tomorrow.

I am going to focus on just figuring out the issue of determining orientation for the scattered nodes. Let me have lunch here."