"10:40am. Let me chill and then I am going to get started. I slept well tonight. Today I really need to figure out what the deal with properly orienting the scattered objects.
10:55am. Let me start.

https://www.youtube.com/results?search_query=houdini+scatter

Let me start by watching the scatter tutorials.

11:15am. Sigh, that first vid on advanced scattering is not giving me what I want. Let me try the scatter node again.

Well, it crashed, but I did manage to get the leaves to work properly just before that happened.

Let me see if I can properly place them on the stalk.

11:40am. Success. Now the leaves are pointing in the right direction.

Though that still leaves the mistery of how to use the scatter and align node properly. I think that just orient should be enough to properly set it up. So why is that not working correctly?

https://youtu.be/mgSGYRXnEmg
Tutorial 6. Scatter Node

Let me watch this thing.

11:45am. I do not get how alignment for the Scatter and Align node works. I think that normal direction should be Y. But what about the forward and target forward direction.

Ok there are only 9 combinations, I can go through all of them.

11:50am. As it turns out, one combo works. Forward direction Z and target forward direction Y gives me the right results.

11:55am. Ah damn it. One of the leaves is oriented incorrectly causing it to stick from the inside.

12pm. This seems to only be an issue with the scatter and align node. Is this annoying? Yes it is. I am really having trouble getting the align part to work correctly.

12:05pm. When I get rid of orient, it starts working again. I have no idea. It might be worth asking around. Let me watch the tutorials. I just need to sit down and do it.

Hrmmm, let me just ask.

12:15pm. Fuck. I can't figure out what settings caused it to stick out from the inside. Right now no matter what I try, I get the correct results.

12:20pm. Sigh, no, I can't find the buggy state.

Well, I suppose that is not bad. Houdini can be buggy. I've had it crash plenty of times, so maybe this orientation bug was something weird.

Let me go back to watching the tutorial. I've watched a few of them already and I still haven't internalized how to instance between different choices. I know a few ways of doing it, but all of those ways are shit.

https://youtu.be/mgSGYRXnEmg?t=1013

Hmmm, I should keep this in mind.

1:05pm. Watched the scatter node tutorial. My focus was low, but I got just a bit from it. That is fine for now.

Independently, I did figure out how to do the flower so that will be my accomplishment for the morning. Step by step, and all that.

> I think honestly studying Houdini sounds like a very valuable investment of your time for a project such as this.

Got this reply to my PL thread. Honestly, I really hope so. I am considering just ditching Houdini given how much of a handful it has been so far, but I should do at least a single scene in it before I decide. If I waste a month then so be it. I am under a lot of pressure to not have that happen though.

1:10pm. What I am going to do next, now that I finally got a single flower is figure out how to turn it into a function. I'd want to pass in the seed as an argument and generate a bunch of different flowers based on that.

1:15pm. How about I leave that for later and get started with the pool scene itself? I've gotten the prereqs out of the way, so I should do some modeling.

Let me have breakfast here first. It is time for the usual B 'n C.

2:50pm. Done with breakfast and chores.

In my current state, chores sure are exhausting. At this rate, I have no idea how long exactly will it take me to overcome this illness. But that does not matter. I need to hunker down and just get on with the art.

Let me load up blender. I need to look up the scene again.

30 x 26 x 6. That is the size of the pool island.

My god, looking at my old stuff, I sure have done a lot. Maybe it is insane to ditch all this old work for just the potential of new one being better.

Let me do it. Let me give it a try.

3:20pm. Oghhhhh, this is like pulling teeth.

3:40pm. Now I am trying to figure out how to align the fence segments.

4:15pm. Had to take a short break. I am trying to remember how I aligned the fence segments in Blender and am utterly failing. Did the normals get carried over from what was originally a face?

Remember those experiments I did when I played around with normals of various meshes in Blender? No. I forgot all about that.

4:20pm. Ok, this is not just my imagination.

Houdini is really hard to deal with.

I made that 30 x 26 box to represent the pool island similarly to in Blender. Then I used the PolyExtrude to inset the edge which I've used to distribute fence segments on.

How do I rip the part of an edge in order to create an entry point? Also the sides have points that are on a curve. In Blender I ripped all the edges apart, scaled them and then distributed the fence segments. In Houdini, how am I supposed to do that.

4:35pm. I am drawn towards using booleans, but that is already a solution I am leaning towards out of convenience rather than actual desire. What I really want to do is rip out those edges and split them up and scale them, the same as in Blender.

4:50pm. You know what Houdini feels like right now?

Like Haskell. And Haskell is one of those languages you talk about rather than actually use.

Houdini is absolutely overwhelming me. Right now, I am at an absolute low point and it feels like it is kicking me when I am down.

I am looking at the Blender scene, thinking about how I would replicate it in Houdini and am groaning internally.

I simply lack the most basic of modeling techniques. I am frozen in indecision here.

And it is not really my fault, but with the way the program was designed. I asked in the /3/ Houdini thread, but moving around the viewport is still a pain in the ass. This is one of the things I haven't gotten comfortable with. It is grating on my nerves.

The reason I am in such a horrid situation is because I could not accept some of Blender's minor limitations and looked elsewhere. Well, I sure am paying for that now.

5:10pm. FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF---

It is just not going for me!

I guess I'll close here for the day. I am absolutely useless right now. I can't do a single thing right.

I feel completely aimless and directionless. Do I want to go this way or that, I can't seem to muster emotion either way. Deep down, the only thing I can dredge up is some primitive aggression.

5:15pm. Honestly, I am disappointed with Houdini's scattering capabilities. Despite all this learning, I can barely remember how it should be done. Not even a single time have I been impressed with its design.

This a sharp contrast to two and a half weeks ago when I felt a lot of power at my disposal in Blender.

5:25pm. This disease is so nasty. It is like all color drained from my life. I wish my father had been more careful not to catch it.

...Let me close here before I write something I will regret. My energy is low, my inspiration is zero. For all intents and purposes, I am an NPC in my current state. I can only hope that I recover from this. If I am still a zombine by the end of the month, I'll forget about Heaven's Key and start looking for a job.

For now, forget about all this hard stuff. Let me rest."