"10am. Almost ready to start. I went to bed at 12am, and have gotten up 30m ago. The disease's effect is definitely waning. Let me read the Yakuza Princess chapter and I will start work on varying the vines.
10:10am. Let me start.

Hmmm, I've been wondering this for a while. What is photogrammetry? I see this being ref'd in the docs everywhere.

> Photogrammetry is the art and science of extracting 3D information from photographs. The process involves taking overlapping photographs of an object, structure, or space, and converting them into 2D or 3D digital models.

10:15am. I ma thinking. Let me just ask.

10:20am.

///

The stalk for this flower is made from a curve. I want to add variation to it, and I am trying to figure out how to interpolate between this and a straight curve.

One way of doing it would be to pick the middle point of the curve and scale it so the curve gets straighter.

But if I do that it might be good to also rescale the curve so it has the original length. How could I do that?

///

Hmmm, there is also something called attribute interpolate. And I need to keep in mind that position is an attribute.

Actually, I won't ask. I figure it out on my own already. Let me just put my plan into action.

11:25am. Damn it, this is so annoying. How do I make one HDA inherit params from another.

12:35pm. I've figured out how to straighten and interpolate long ago using blend shapes. I've also figured out that it is just possible to drag the node into the parameters. Houdini is great here.

But right now I am really wracking my brain as to how I could vary the individual flowers.

12:45pm. Decided to actually ask. Copy stamping? But how would that work when it is not even in the same scope.

1pm. https://www.youtube.com/watch?v=-tTzZZuoRco
[VEX for Algorithmic Design] E16 _ Geometry Functions

I am stuck mentally on this issue. As a programmer I understand that the type of the HDA is wrong. I have `geo * geo * geo`, but what it should be is `geo * (point -> geo)`. Maybe I could hack it if I could pass VEX scripts via channel functions.

1:15pm. Focus me. I am way too lost in thought. In the worst case, I could just inline by hand.

I am just wasting time here. Let me have breakfast."