"10:55am. I am up. Let me chill just for a bit more and then I will start.
11:05am. The latest Konchuki chapter was wild. Let me start.

11:25am. No, what I had thought of yesterday would be too much of a pain in the ass to do. Forget it.

Right now, I do not have a good way of getting a tuple of outputs from a for loop. Let me ask about that.

11:45am. I asked, but I had a different idea from yesterday. What I am doing now is packing the body inside Flower Body. It is a HDA helper. Yesterday and during the night I've been thinking about producing multiple outputs, but that is wrong. What I need to do instead is do more packing and delineate them with groups. That will allow me to produce multiple outputs succinctly.

I need to make more liberal use of packing. When I merge packed objects, the underlying refs do not need to have the same fields. Meaning I do not need to have the bulbs have the `curveu` attribute in order to match the stalk.

12:55pm. Ok, I have it. I think I am finally satisfied with the level of control. I did a great deal of refactoring. I wasn't wrong about producing groups after all. But what I should not have neglected is packing. It is useful for more than just saving memory. It is an indispesable organizational tool. I also need to keep in mind that I need to mark object with groups so I can split them after merging.

This is the way to go.

1pm. Right now the whole scene has as much detail as the original, but it bakes in a snap.

1:10pm. Time for that break. I've started slacking anyway. I am perfectly satisfied with this.

2:25pm. Done with breakfast and chores. Let me finish the ep of Kenja no Deshi and I will resume. I think it is finally time I put down that bench and did the props.

2:45pm. Let me resume. Let me make a reply in the Houdini thread. Then I'll get started on the props. I've been working on these vines for way too long.

3:10pm. Did the update in the thread. I feel like a little break. Let me make up my mind to do the props.

3:40pm. Let me finally start the afternoon session for real. I want to make just one final adjustment to the flowers.

4:20pm. Ok, I've figured out how to orient the leaves along the stalk curve as opposed to in the upwards direction. The normal way is good for branches as they fall in the direction of gravity, but for the small leaves I really wanted them to go in the direction of the mesh curve.

It turns out the scatter and align can take in the target forward as an attribute. What I need to do is simply pass the `up` attribute there. I had to change how the flower HDA works to get it out, but now it works just fine. Great!

4:25pm. By giving an output node to an object, you can make sure it always displays the end result.

It is really good I figured this out. That is another bit of knowledge added to my mental library. It is important that I know how to pipeline the curve data up to the mesh.

Now...

4:35pm. I need to get started on the props. I literally have everything I'd ever need with the plants. This was the third refactor in the last half week, and I even put in a bit extra just now.

I've got it down.

The trouble is that I am backing in my own glory instead of moving on to the next step.

Let me just put down that bench.

4:40pm. Done. That is a good bit of icebreaking. Now let me do the sunbed. Let me open Blender so I can get a sense for its dimensions.

4:50pm. This is killing me. Drawing curves is beyond painful.

What I need to do here is rewatch those hipflask tutorials on construction planes. I should have done this earlier while doing the entry posts. Instead of modeling those using extrusions, I should have just done a curve and then swept it, but I had so much difficulty putting it where I'd want to.

https://www.hipflask.how/products/houdini-made-easy-01-the-core-essentials/categories/1273889/posts/4428289

Let me start from here.

5pm. > To enter the construction plane manipulation mode, you just tab the slash key.

Ahhhhhh...

5:05pm. Hrmmm, how do I manipulate breakpoints directly. It is no good if I have to fiddle with regular points.

https://youtu.be/SNDJzhOZZeg?list=PL4XNGOsVprEoijbvM0eLlwihBx0M2p1Lp
The Weekly Houdini Tutorial: The Anatomy of Nurbs Curves

Let me watch this again.

https://youtu.be/SNDJzhOZZeg?list=PL4XNGOsVprEoijbvM0eLlwihBx0M2p1Lp&t=38

I see a method there in the pane. That is what I should manipulate.

5:15pm. Hmmm, I could just draw a Bezier curve instead.

5:25pm. Ran into some ridiculous recursion related bug.

Sigh this will take a while. It turns out I have no idea how to draw curve. I know how to do it in Adobe Illustrator and Clip Studion. Even in Blender I should be able to manage it. But here things are so confusing.

5:40pm. I finally put the 4 sunbed points in the right place. It turns out that order 3 Bezier curves are hard to deal with. Let me watch the tutorials all the way through.

5:55pm. https://youtu.be/iv51A4syohU?list=PL4XNGOsVprEoijbvM0eLlwihBx0M2p1Lp&t=31

I really need to figure out how to draw using breakpoints.

6pm. I did not manage to figure out how to draw breakpoints, but the path tool is amazing. It allows me to draw nurbs curves, but also to control and rotate CVs.

6:35pm. As amazing as it is, putting an object network in geometry causes the outside geometry to disappear. As a rasult I lose track of relative sizes. Sigh. The reason why I need to do that is because the path tool just tosses CVs as objects on the top level.

I think I'll just stick with Bezier curves.

7:30pm. Let me go have lunch.

Actually, it will be in a bit.

8pm. Done with lunch. Let me close here.

Right now I am trying to make that sunbed. You wouldn't think this would be challenging for me, but you run into things.

First of all, I need to investigate what Bezier surfaces are.

I do not know how to make effective use of them. I fiddled around for a bit, and then I just resampled after that.

The second, I need more basic modeling skill in Blender. With the lantern plant, I studied a lot how to scatter and instantiate, but I still have great weakness in basic modeling. I still do not know what 99.9% of the nodes in Houdini do. A sunbed is possibly the easiest thing imaginable to model, but already I am running into limitations of my skill. In Blender it would be much easier. I can think of ways of doing it there. But here it is hard.

8:15pm. Also, I find myself being annoyed at not being able to go in isolation mode in Houdini. In Blender I just have to press slash to isolate an object, there should be something similar in Houdini, and I should really find out what it is.

The challenge of doing modeling is knowing which technique to use in what situation.

8:20pm. I've been in the zone the whole day today, but I think I should close here. Tomorrow I'll launch an in depth investigation of things I've overlooked.

I need more basic modeling skills to do these props effectively in Houdini.

8:25pm. I want to do them in it too, even though at this point I have the option of importing Blender stuff. But still, picking the best tool for the job requires me to actually understand the tools. Unless I put in the effort to get to know what is unfamiliar, that will never happen.

Forget the time taken. At some point I'll have the skills to move smoothly in Houdini. Until then I should keep practicing.

Eventually my 3d art skill will reach a sufficient level to work on my own things unimpeded.

8:30pm. Right now I might have completely failed my ML quest, but put it another way, all I need to break through to the next level is get an AI chip and implement a genetic programming system. The Singularity is quite near, yet so far away for me. Just when will I get rewarded for my efforts? I guess this journey is its own reward."