"11:30am. The gaming sessions are too good, and I am getting up late again. Sigh. I guess I'll skip the morning session. Let me have breakfast here. Then I'll do the chores. After that I'll figure out how to do basic modeling in Houdini. I had time to think about it.
1pm. Done with chores and breakfast. It is time to start. I need to get my Houdini basic modeling skills up to par. The things I learned in the Hipflask course are absolutely vital for that. If it wasn't for it, I'd be hosed right now. I need to do these props so I can move to texturing. I actually have no exp in that, and it will be a challenge I need to surmount.

1:20pm. I have a good feeling as I do my modifications.

I actually did one of the two supporting beams for the sunbed. It is positioned properly. And right now I am thinking how to follow up on that. The way I did it, I can't just copy it to the other end.

> Edit has an optional second input for reference geometry. When the second input is connected, the Edit node transforms your edits relative to the differences between the geometry to edit and the reference geometry.

Right now I am reading the docs for the edit node. I need to think about it. The edit node has so much stuff in it, and I need to get familiar with it. The edit node is in fact the equivalent of edit mode in Blender.

Now what is this ref geometry. I really need somebody to show me more in depth. Words will only get you so far.

1:40pm. The Edit node has two examples. I need to take a look at them.

Hmmm, I just realized that the examples have object networks inside object networks. But I can't find a way to do that using tab. I wonder how that could be done?

1:50pm. The first example is so ambiguous. It is supposed to demonstrate editing with ref geometry, but even if I cut the wire, I still get the same result.

The other example is literally just a plain grid. Houdini does have some 'just doing our job' style corporate laziness baked into its design.

I really wish I could access the rest of the Hipflask vids for free.

...I decided to make a request in the piracy thread, but I doubt I'll get anything.

Let me leave that aside. Let me redo the beam from the start.

https://www.sidefx.com/docs/houdini/nodes/obj/viewportisolator.html
> The Viewport Isolator is similar to the viewport’s View Mode display option that lets you ghost or hide other objects, but it doesn’t restrict you by the OBJ hierarchy.

I'll have to check out that view mode. Nevermind it for now.

2:40pm. https://www.sidefx.com/forum/topic/5907/
Q: Edit SOP's Reference geometry input does what exactly?

I'll have to figure this out.

Right now I am thinking about the beam. Yes, I did it, but I did not do it exactly right. The sharing is not there.

https://youtu.be/GHyZVT6t0Z4?t=361

The only thing I am wandering about is the normals are usually a vertex attribute, but here they are tagged to points. How does that work?

3:15pm. I am thinking of dropping Houdini again. This is so annoying.

3:50pm. Let me take a short break here. I figured out how to make edits local. The second input to the node is key after all.

5pm. It is true that the break was overly long, but I am still thinking about it.

No, the way I've done the beams is not right. Damn.

5:15pm. https://www.sidefx.com/docs/houdini/basics/objects.html
> This is not to say that transforming at the geometry level is bad: often you need an object’s shape to deform over time. However, you should remember to look for opportunities to accomplish your goals by transforming objects before you look at transforming surfaces, because the former is more efficient.

Fuck. I made a serious mistake!

5:50pm. I am so obsessed with this, that I haven't had time to do actual modeling.

Throughout this whole scene, I've made a big blunder! I've been transforming things inside objects everywhere.

Here today with this sunbed I've been bashing my head how I can do my modeling exactly in place only to realize that the best way to do that is to just use objects.

Is there a way to have the construction plane match the object's axis? I'll have to look into that.

6:10pm. Done with lunch.

7:25pm. Let me just close here. I did the sunbed and thne spent the whole day in thought. This totally reminds me of programming. Today I think I completely internalized the uses of construction planes. I do not think that those curves will be giving me any more trouble.

I also noticed that align tool in the modify shelf.

7:35pm. Agh, I am still thinking about it. To think I'd spend an entire day brainstorming just how to properly set objects.

But it is important part of the modeling technique. The fact that I could not do the entryway posts using curves was quite sad. I really needed to dwell for a bit on what I learned in the Hipflask course.

With this, I'll deal with the towel rack tomorrow. Well, the rack itself will be easy, but what about the towels? I guess I should learn a bit about cloth simulation in Houdini.

But first thing's first. Let me do the rack itself, then the pool nibbles, then a part of the entriway walkway. I'll import that other fence segment.

Let me just do this stuff. I need to get it out of the way. Just how much am I going to spend on this scene?

7:45pm. Tomorrow I am going to absolutely do the rack, do the pool, do the other fence, do the menu. The pool will also need a ladder. I'll take care of that as wall. I'll just turn a grid into lines, blast the top and bottom away, bend the top so it curves downwards.

Do I need anything other than that? I think no. So after that I'll be able to look into texturing.

Texturing will take me a while, but after that will come Luna. Once I complete this I will have a full 3d skillset. A few more scenes of same complexity as this and I can consider myself in the upper 2/5 range. I'll only need experience to break into the pro realm properly after that.

Houdini is so annoying, but I am going to master it. Nothing is going to stop me in this.

After that I'll am going to have full speed, and will be able to create scenes worthy of Heaven's Key.

I am going to cultivate this to its limit, and then make a real attempt at producing something worthy of admiration, if not value. As usual, I'll take the hardest route to that 2k for the AI chip. But the hard way is usually the most satisfying.

Won't it be wonderful to actually do something of value during the day instead of just trying to internalize the basics. A few more months and I will definitely reach such a level.

Even though I haven't done much, I've been in the zone the entire day. So that is a sign that I've learned."